코드정리

+++ src/main/java/com/fitmate/oauth/controller/UserController.java
@@ -1,6 +1,7 @@
 package com.fitmate.oauth.controller;
 
 import com.fitmate.oauth.dto.ResultDto;
+import com.fitmate.oauth.service.TokenService;
 import com.fitmate.oauth.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpHeaders;
@@ -15,14 +16,21 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
 
     private final UserService userService;
+    private final TokenService tokenService;
 
-    public UserController(UserService userService) {
+    public UserController(UserService userService, TokenService tokenService) {
         this.userService = userService;
+        this.tokenService = tokenService;
     }
 
     @PostMapping("/user/nickname")
     public ResponseEntity<ResultDto> updateNickname(@RequestParam long userId, @RequestParam String nickName, HttpServletRequest httpServletRequest) {
         String accessToken = httpServletRequest.getHeader(HttpHeaders.AUTHORIZATION);
+        if(!tokenService.isTokenValid(accessToken)) {
+            return ResponseEntity
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .body(ResultDto.of(401, "UNAUTHORIZED. check accessToken"));
+        }
 
         if(userService.updateUserNickname(userId, nickName)) {
             return ResponseEntity.ok(ResultDto.success());
@@ -34,7 +42,14 @@ public class UserController {
     }
 
     @DeleteMapping("/user")
-    public ResponseEntity<ResultDto> deleteUser(@RequestParam long userId) {
+    public ResponseEntity<ResultDto> deleteUser(@RequestParam long userId, HttpServletRequest httpServletRequest) {
+        String accessToken = httpServletRequest.getHeader(HttpHeaders.AUTHORIZATION);
+        if(!tokenService.isTokenValid(accessToken)) {
+            return ResponseEntity
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .body(ResultDto.of(401, "UNAUTHORIZED. check accessToken"));
+        }
+
         if(userService.deleteUser(userId)) {
             return ResponseEntity.ok(ResultDto.success());
         } else {
